Code style refactor

+++ Copper-Engine/src/Engine/Core/Core.h
@@ -46,7 +46,7 @@ namespace Copper {
     typedef int                 int32;
     typedef unsigned int        uint32;
 
-    typedef long int            int64;
-    typedef unsigned long int   uint64;
+    typedef long long           int64;
+    typedef unsigned long long  uint64;
     
 }
